despliegue-de-algoritmos-main/solutions/twitter-sentiment-streaming/twitter-to-pubsub/requirements.txt
tweepy==3.9.0
loguru==0.5.3
google-cloud-pubsub==2.1.0